refactor: move totens list to admin screen
